rename clients package
