+++ src/test/java/org/folio/edge/oaipmh/utils/OaiPmhOkapiClientFactoryTest.java
@@ -5,8 +5,8 @@ import static org.junit.Assert.assertNotNull;
 
 import io.vertx.core.Vertx;
 
-import org.folio.edge.oaipmh.clients.aoipmh.OaiPmhOkapiClient;
-import org.folio.edge.oaipmh.clients.aoipmh.OaiPmhOkapiClientFactory;
+import org.folio.edge.oaipmh.clients.OaiPmhOkapiClient;
+import org.folio.edge.oaipmh.clients.OaiPmhOkapiClientFactory;
 import org.junit.Before;
 import org.junit.Test;
