+++ src/test/java/org/folio/edge/oaipmh/utils/OaiPmhOkapiClientTest.java
@@ -14,10 +14,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.http.HttpStatus;
-import org.apache.log4j.Logger;
 import org.folio.edge.core.utils.test.TestUtils;
-import org.folio.edge.oaipmh.clients.aoipmh.OaiPmhOkapiClient;
-import org.folio.edge.oaipmh.clients.aoipmh.OaiPmhOkapiClientFactory;
+import org.folio.edge.oaipmh.clients.OaiPmhOkapiClient;
+import org.folio.edge.oaipmh.clients.OaiPmhOkapiClientFactory;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
